pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skylattice"
version = "0.1.0"
description = "Private, evolvable personal agent foundation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8,<9"]

[project.scripts]
skylattice = "skylattice.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
